requirements.txt
fastapi==0.129.2
uvicorn[standard]==0.41.0
streamlit==1.54.0
sentence-transformers==5.2.3
google-genai
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.3
numpy
scikit-learn
accelerate
datasets
pg8000
cloud-sql-python-connector[pg8000]